Update issue 344
It seems that at gnome-canvas-shape.c:1441 gdk_pixmap_new is called as a result of ctx's width and height not being big enough. However, the ctx's height and width aren't updated and the ctx is reused in subsequent calls, resulting in the repeated allocation of ctx->mask.
Attached are very rough notes from my debugging.

Everything is actually in r1240, but there was a typo in the issue number.