Update issue #410 Get the basic code working.  Options are not passed on at the moment, but the tree is plotting correctly.
The test only checks for eval error at the moment.  We need to read the file back in and do some cross-checking, i.e. for correct number of nodes, lengths internal/named status etc.

Also add comment type to the parameters metedata.  This allows short messages to users in the options where it is best to be very clear, rather than out of the way in tooltips.