Update issue #408
Added methods to read shapefiles, appears to be working on examples tested.  There is a bit of an issue with setting CELL_SIZES, the value appears to be stored redundantly under Groups, so when the value in the basedata_ref of the shapefile data is updated, it doesn't update the value used for displaying (which remains at the default).  I changed the default to 100000 just to test the coastline shapefile.

Handling labels is done a bit differently to column data, although the code is reused, i've made it lenient about choosing a label column, to give the option of using a constant label (1), eg to use data simply with X,Y 'columns'.  It should handle Z,M and allow arbitrary groups/columns.

Added PPD for GDAL on windows, compiled using MinGW 64 and build against Strawberry Perl (5.16).  This was extremely difficult to get it to build, however a re-build on the 32 bit toolchain (and 32bt strawberry perl) should be possible, and it builds with 5.18 as well.  Its probably worth adding the modified source to the repository to allow rebuilding.