Update issue #408 Get the shapefile field names in file order. Give the shapefile coords distinct names so they don't clash with any field names.  The leading colon is an invalid field character so ensures distinctness.
Other issues remaining:
The shapefile import seems to be using column names, whereas the text import uses column indexes.  Names are a useful thing, but we should map column names to column ids to keep things consistent.  In the shapefile case this gets a bit hairy, as the x, y, z & m will be at the front of the array, so any scripting interfaces will need to check for the z & m fields.  Or we can only enable them when asked.