Update issue #295
adding methods to handle common progress bar window.  it shows multiple progress bars active in an extended vbox, and removes each as use of the progress object finishes.  this makes use of the existing ProgressDialog object and interface, so no change is needed to calling methods.

it may need changes eg for methods such as randomisation, where multiple very short progress instances are called.  the progress dialog appears and disappears quickly as each is brought up and removed in succession, its not clear what the preferred behaviour is, perhaps changes to the display of progress in these instances could be used.  a possible alternative change is to maintain completed entries, in a scrolling window with new entries at top.