Update issue #295
Handling display of progress dialogs, eg catching when the window is closed, and keeping track of all relevant dialogs attached to progress.  It handles the basic case, but there seem to be issues eg after closing a progress window the behaviour is different when run a second time.  Also often display of first progress (eg "1 of 101 this run") doesn't show, not clear why.