Update issue #295
fixed issue with display of progress not updating first display, as a result of time interval test preventing update of labels

fix issue with progress window not displaying after window is closed.  changed so window is allowed to be destroyed, $self->progress_bars is cleared, all progress dialogs are removed and a new init call is used when progress is next called.