Update issue 378 Change the flag variable and widget names for one basedata per imported file so it matches its purpose more closely.  Move its call up in the file as dlgxml has changed by the time it was being called, so the widget no longer existed.  The mutable $dlgxml aspect of the code needs to be changed methinks, as it is confusing. Make sure we use array refs when setting cell size and origin parameters. Clear out the swapped="no" calls from the glade file.  These will no doubt creep back in due to the glade editor so constant vigilance is needed.  perhaps a test should be added, and maybe a commit hook to remove them.
Next is to make sure we add all the files.  Iterating over a list is perhaps the simplest approach since the arguments are the same except the file names.  In the single file case it can just be a single item list, or we can pass an array ref as the first list item when we want to lump the files into one basedata.