Update issue #378
Modified import to allow import of multiple files, with the option of loading as new datasets, or searching for existing basedatas with the same name (determined from file), and adding to the existing dataset.

Operation is controlled by the "Multiple as separate.." and "new" checkboxes on import.  If multiple and new chosen, new separate BDs are created, even if existing are present.  If multiple chosen and not new, for each imported file it will search for an existing BD with the same name, and add to it, however if an existing is not found, a new one will be created.