Add a number of speedups.
Matrix/Base.pm:  Add new method get_value_or_undef() which is considerably faster for the common use case of get_value, as it does not need to check for existence first, and does not worry about self-similarity (which is rarely used).

Tree.pm: Reduce the search space when looking for common ancestors, taking advantage of the depth properties of the path lists (each element is one depth lower than the next, which is its parent).
  
The rest are very minor variable shifting to reduce needless creation, or bulk creation of variables.