Update issue #353
Changed a number of files in trunk to unix format, so they are the same as those in the 353 branch.  This allows merging without overriding the changes in the 353 branch (if you do unix2dos on those files in 353, the changes from the common ancestor aren't recognisable and it becomes hard to merge).