Update issue #353
Merging trunk r2524.  Blocks of glade file that have significant updates in 353 were left as 353 changes, eg most of wndTabs, notebook, spatial windows, with the assumption that the changes in 353 are the most recent.

Files with dos new lines were modified to same format in trunk, to allow revisions in 353 since common ancestor to be identified.  (move back to common format after)