Fix some test failures exposed when testing under perl 5.20
The main one is that the region grower tests weren't using link_recalculate, so add methods to make sure this is the default linkage used (should make it the only possible option, as per the GUI).
Also clean out remnant matrix elements when the no_clone_matrices option is specified.  Region growers had leftovers when clustering stopped early and this triggered matrix recycling on later analyses.  
Add a method to Matrix.pm to do this clean out.