Update issue 510 Rework the csv sep_char guessing process.
If a second line is passed then it will cross-check which of the separators gives the same number of fields and return that.
Also rework the order of checking in Biodiverse::GUI::BasedataImport, plus copy more of the Biodiverse::BaseData logic.  This should be refactored since so much is common.

Also add tests.