Update issue #538 - revise the help URLs.
Point help URLs to the purl.org/biodiverse versions, now that they redirect to github instead of google code.  The exception is the versions check, which currently points to the raw github file.  We might want to fix that.

Also update the authors list in the about dialogue, and make sure its link goes to the correct place.

Signed-off-by: Shawn Laffan <shawnlaffan@gmail.com>