Avoid storing descendant node refs in the cache.
Turn off caching in get_all_descendants.

Add new methods get_names_of_all_descendants and get_names_of_all_descendants_and_self to access hashes of the names of the descendant nodes.

Names are all that is needed in nearly all use cases, not node refs.  The values stored in the hash are the number of children of each node, but that could change in the future.  It is currently only the GUI click-popups which make use of the values.

Not storing node refs avoids the ludicrously deep recursion which was evident for one example tree, causing crashes on Windows under Strawberry perl when cloning or saving.

Updates issue #549

Signed-off-by: Shawn Laffan <shawnlaffan@gmail.com>