Search a subset of index offsets when a spatial condition sets index_max_dist.
Currently this still returns early when the distance is less than twice the minimum index resolution.  That ratio needs to be tuned, perhaps.

If we do the subset search then we also now return the subset of offsets if the search results in more offsets than the subset.

There is still old code from long ago, before the shortcut was added, which does effectively the same thing.  It can be cleaned up next.

Updates issue #550

Signed-off-by: Shawn Laffan <shawnlaffan@gmail.com>