Add an extra test to ensure the set of neighbours matches, not just the count.
This is implemented on the assumption that the non-indexed search is the canonical result, and we then check the rest match.  We should set up some gold standards as we do for the Indices.

Also set the prng_seed in the sp_block_select tests, as the new check was otherwise failing.

Signed-off-by: Shawn Laffan <shawnlaffan@gmail.com>