Don't check the index unless it will result in fewer loop iterations overall.
This assumes the neighbour comparison checks cost as much as the index offset checks.  In some cases this won't apply, but these are not expected to be common.

Implementation required the addition of new subs to Biodiverse::Index to calculate the mean number of items per index element.

Updates issue #545

Signed-off-by: Shawn Laffan <shawnlaffan@gmail.com>