Some speedups and cleanups for the subset code.
We now store and re-use the spatial output used to define the subsets.  This way we avoid recalculating it.  At some later date we might need to only do so if the results are invariant, but we don't have methods to check for that yet.

Also make use of Biodiverse::BaseData::add_elements_collated to reduce some repeated code.  More refactoring of those blocks is in order.

Updates issue #554

Signed-off-by: Shawn Laffan <shawnlaffan@gmail.com>