Add spatial_condition widget to the randomisation metadata and thus GUI tab.
Need to handle cases where no spatial condition is specified, but a def query is.  Possibly the approach to use is to add a sp_pass_def_query method to Biodiverse::SpatialConditions, and use it as the default when no spatial conditions are passed to Biodiverse::Spatial::run_analysis.

Updates issue #554

Signed-off-by: Shawn Laffan <shawnlaffan@gmail.com>