Add capability to hold some set of labels constant in a randomisation.
Extras include:

1.  The addition of a text widget to the parameters, albeit it needs to use similar tricks to the spatial conditions widgets so it does not take up too much real estate.

2.  A new method to add collated elements - add_elements_collated_by_label, indexed by label. Need to rename add_elements_collated to be explicit about groups.

The listing of labels in the GUI uses one label per line, which will cause problems if labels contain newlines.  No such cases are known yet, but it will probably crop up as an issue in the future.

Also remove one todo test in 28-Randomisation.t, as it is covered by an existing test.

Updates issue #556

Signed-off-by: Shawn Laffan <shawnlaffan@gmail.com>