Track the number of iterations associated with each PRNG init state.
This will allow replication of results, as we know the initial state and number of iterations to run for each separate process.

The sum of these would need to be subtracted from the total iterations parameter, as it includes any iterations run on the "master" basedata, plus all those which have been reintegrated.

Add a test to ensure we don't reintegrate if we have already run using a specific PRNG stream (i.e. the init state is in the list of those we have already integrated).  We skip even if the number of iterations is larger, as we have no way of knowing which specific comparison results should be skipped.

Updates issue #558

Signed-off-by: Shawn Laffan <shawnlaffan@gmail.com>