Add calculation for phylogenetic abundance.
The abundance of each branch is the sum of its named descendant abundance, multiplied by the branch length.

This is on the basis that PD is feature diversity thus if a branch is 10 units long then it represents 10 features.  If that branch has 5 named descendants, each with unit abundance (=1) then there are 10 * 5 features from that branch in the sample.

Updates issue #559

Signed-off-by: Shawn Laffan <shawnlaffan@gmail.com>