Squeeze some more performance out of the last-if merge.
If we re-use an SV then we avoid generating new ones all the time, at the cost of all added vals being refs to the same SV and thus a change to one will change all.  Still, we are assuming the values will be set later, or don't matter.

Updates issue #560

Signed-off-by: Shawn Laffan <shawnlaffan@gmail.com>