Try optimising the b and c branch search, unsuccessfully.
This approach iterates through the children of any b or c branch to avoid the neighbour search.  However, it costs as much as the neighbour search so comment it out for now.  If we can amortise the search for branches within the neighbour sets across multiple indices then it can perhaps be reinstated.

Updates issue #577

Signed-off-by: Shawn Laffan <shawnlaffan@gmail.com>