Add LICENSE file.
This PR is part of the CPAN Pull Request Challenge 2015 for the month of
February.

This commit an attempt to fix the following warning noticed on [1]:

`has human readable license`

`This distribution does not have a license defined in the documentation or in a
file called LICENSE

Add a section called "LICENSE" to the documentation, or add a file named
LICENSE to the distribution.`

Biodiverse being released under the GPLv3 license, this is a verbatim copy of
the GPLv3 license.

[1] http://cpants.cpanauthors.org/dist/Task-Biodiverse-NoGUI