refactor get_significance_from_comp_results
Hard code the default thresholds to avoid repeated map calls.

Only calculate the two tail scores if the relevant one tail threshold is passed.

Also test that we get at least some defined results (although this does not happen if we up the number of iterations to 99 due to the data set used).

Updates #607