save to storable if loaded from Storable
If a data set is loaded from the storable format
then we make sure to use that when saving,
unless told otherwise by the method arg.

This should allow better (but not perfect) back-compat if users are
switching between old and newer versions.

Any new data sets are saved to sereal by default.

Updates #358