add method to increment the multi-select colour
Also add methods to get the colour palettes as Gtk2::Gdk::Color objects.

The increment process needs some thought yet,
as it currently resets to red if a colour from
outside the palette is used.

This will be simplified once we keep track of which nodes are which colour.

Updates #600