swap_to_reach_richness_targets: use a different tracker list structure
The List::Unique::DeterministicOrder approach is faster
than binary searches on sorted lists.

Determine using profiling with a data set with many groups
and labels with large ranges so there is a lot of swapping.

This is a breaking change as the randomisation order
changes from previous versions.
Some tests need to be updated.