optimise RichnessEstimation.pm
Faster::Maths allows a substantial speed-up (7-30%)
of arithmetic code, of which there is a lot in this module.

It currently segfaults in some circumstances so
lexically disable it as needed.
This has been reported to its RT queue.

Also avoid some looping and memory usage in
the ACE and ICE variance calculations.