Indices: calc_abc adapts cached calc_abc2 or calc_abc3 results
If calc_abc2 or calc_abc3 have already been calculated
then we can just grab their results and set the label hashes
to have values of 1.

This avoids a lot of looping in several circumstances,
for example where large neighbour sets in a spatial
analysis or where cluster indices are calculated per node.